refactor(otel): extract telemetry setup out of app.py into a package
app.py was 577 lines, almost all OTel provider setup, four instrumentation
guard blocks, a FileCache monkeypatch, and a logging bootstrap, with only ~6
lines being the actual WSGI app. Split everything but app construction, CORS,
and two telemetry calls into src/telemetry/ so app.py reads as what it is.

- telemetry/_logging.py: leaf module (log bootstrap, otel_log logger)
- telemetry/instrumentation.py: redis/sql/boto/http guard blocks, verbatim
- telemetry/mapproxy_cache.py: FileCache tile-tracing patch, verbatim
- telemetry/__init__.py: providers, install_instrumentation(), and
  init_when_safe() (the postfork/worker_id dispatch), now idempotent

Wire the new package into the image and CI: Dockerfile COPYs src/telemetry/
alongside app.py, and pull_request.yaml gets a python-smoke-check job
(py_compile + a real `import app` against a throwaway MapProxy config) so a
broken import fails CI before merge.

Lift-and-shift only, no behavior change: verified the built image's startup
log sequence is identical to the pre-split baseline under a real uWSGI
master+2-worker boot (one provider-init per worker, no export errors).
