+++ Dockerfile
@@ -190,6 +190,7 @@ WORKDIR /mapproxy
 
 # Copy application code and entrypoint (single layer, correct ownership)
 COPY --chown=mapproxy:mapproxy src/app.py       /mapproxy/app.py
+COPY --chown=mapproxy:mapproxy src/telemetry/   /mapproxy/telemetry/
 COPY --chown=mapproxy:mapproxy entrypoint.sh    /mapproxy/entrypoint.sh
 
 # Fix permissions: mapproxy owns /mapproxy, group 0 can read/write (OpenShift)
